refactor: move OpenAPI datatypes to openapi module
